backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
pymysql
werkzeug
google-generativeai
openpyxl
xlrd
PyJWT
python-docx
sympy
antlr4-python3-runtime
pydantic
matplotlib
